baseline-salt*: Fix some bootstrapping issues
We were picking different versions of Salt as we were
using 'latest' in the bootstrap scripts, but (older)
versioned releases in the pillar.

In the Salt Minion bootstrapper, we invoked salt instead
of salt-call.  In the same light, the Salt Master
bootstrapper had an issue because salt never returns a
non-zero error code on auth/connection failure, while
salt-call does.

Lastly, depending on how the script was run, /usr/sbin
might not have been in the path, and thus if[down,up] may
not have been invocable, causing the script to fail.

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>