create-install-media: Copy in the repository
To automate bootstrapping, copy the repository (with the
pillar/secrets, customizations, etc.) into the installation
media, so that preseeds can leverage it if they wish.

For preseeds which don't make use of the repository,
the repository will still be cloned from the internet and
secrets generated on the fly as before.

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>