Add a 'dnsmasq' state
Setup a local dnsmasq server on each node that forwards
requests for the consul domain off to Consul, and include
Consul's service and node zones in the default search
path.

Also, cleanup Consul's "special" requirements at the top
of the salt top file. We should not need a HTTP(S) server
for Consul and can just expose the webserver port on the
management interface instead (as it's protected with a
client cert/key and SSL).

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>