consul: Distribute CLI certs to all nodes
In order to be able to do things like reload `consul`,
(and interact with the local port), we need a CLI key to
be present on the node.

This provisions such a key for each client.

This commit also bumps the raft_protocol and protocol to
version 3, which appeared in Consul 1.10.x.

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>