mariadb-server: Add Galera coordinator/monitor
First cut: coordinator does not have a TTL, so e.g. if it
crashes or fails to push status to Consul for some reason
and MariaDB fails, then state is not properly reflected

However, when things work, the coordinator provides rapid
failure detection at extremely low cost, as only the
leader is polled at a high rate using a kept-open MySQL
client handle, which is backed by a UNIX socket.

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>