hover: Add workaround MFA login requirement
We formerly used username/password authentication with
Hover's API, but MFA is now a hard requirement. Obviously,
this is problematic from an API standpoint.

Fortunately, it looks like the cookies that Hover provides
once completing an MFA login session are good for a year.
So, just keep reusing that session to sidestep the MFA
requirement of sorts...

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>