feat: Debian Bookworm
Add support for Debian Bookworm.

Note that this commit is *honking* for two reasons: first,
because SaltStack no longer natively packages Salt for
Debian, and it was dropped from the Debian repsitory. Thus,
we must package it ourselves or be confined to the obtuse
"onedir" approach that Salt upstrem has decided upon.

Secondly, ability to interoperate with a local Debian
mirror was added. In order to reduce load on the public
mirrors.

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>